feat(editor): export modal replacing File > Export dropdown (#1810)
Extract the GLB/AR-GLB export routine from AppMenu into a shared
lib (src/editor/lib/exportUtils.js) and add an ExportModal with
per-format cards: GLB (Pro-gated with badge + paywall for free
users, AR Ready option) and .3dstreet.json (free). File > Export...
now opens the modal. PostHog: export_modal_opened,
export_modal_export_clicked.
