+++ src/editor/components/modals/ExportModal/index.js
@@ -1,0 +1,1 @@
+export { ExportModal } from './ExportModal.component.jsx';
