+++ src/editor/i18n/locales/.translation-manifest.json
@@ -8,13 +8,11 @@
     "actionBar.zoomOut": "Zoom Out",
     "addLayer.onlyFirstFileAdded": "Only the first file was added. Drop one file at a time.",
     "addLayer.unsupportedFileType": "Unsupported file type: {fileName}. Supported formats: GLB, GLTF, JPG, PNG, WebP, AVIF, PLY, SPLAT, SPZ.",
-    "appMenu.export.arReadyGlb": "AR Ready GLB",
-    "appMenu.export.glb": "GLB glTF",
     "appMenu.export.gltfError": "Error while trying to save glTF file. Error: {error}",
     "appMenu.export.gltfSuccess": "3DStreet scene exported as glTF file.",
     "appMenu.export.uvTransformSkipped": "UV transformation skipped - using original export",
     "appMenu.file": "File",
-    "appMenu.file.export": "Export",
+    "appMenu.file.export": "Export...",
     "appMenu.file.import": "Import...",
     "appMenu.file.new": "New...",
     "appMenu.file.open": "Open...",
@@ -592,7 +590,13 @@
     "segmentSidebar.slopeStart": "Start Edge (m)",
     "segmentSidebar.slopeEnd": "End Edge (m)",
     "managedStreetSidebar.reloadFromSource": "Reload from {source}",
-    "managedStreetSidebar.reloadConfirm": "Reload this street from {source}? Local segment edits will be lost."
+    "managedStreetSidebar.reloadConfirm": "Reload this street from {source}? Local segment edits will be lost.",
+    "exportModal.title": "Export Scene",
+    "exportModal.glbTitle": "GLB 3D Model",
+    "exportModal.glbDescription": "Download this scene as a .glb 3D model file for use in Blender, Unreal, and other 3D tools.",
+    "exportModal.arReadyLabel": "AR Ready: omit people, vehicles and geospatial layers for augmented reality apps",
+    "exportModal.jsonDescription": "Download this scene as a .3dstreet.json file to back up your work or import into another 3DStreet account.",
+    "exportModal.download": "Download"
   },
   "fr": {
     "actionBar.handTool": "Hand Tool (h) - pan and rotate the view without selecting objects",
@@ -603,13 +607,11 @@
     "actionBar.zoomOut": "Zoom Out",
     "addLayer.onlyFirstFileAdded": "Only the first file was added. Drop one file at a time.",
     "addLayer.unsupportedFileType": "Unsupported file type: {fileName}. Supported formats: GLB, GLTF, JPG, PNG, WebP, AVIF, PLY, SPLAT, SPZ.",
-    "appMenu.export.arReadyGlb": "AR Ready GLB",
-    "appMenu.export.glb": "GLB glTF",
     "appMenu.export.gltfError": "Error while trying to save glTF file. Error: {error}",
     "appMenu.export.gltfSuccess": "3DStreet scene exported as glTF file.",
     "appMenu.export.uvTransformSkipped": "UV transformation skipped - using original export",
     "appMenu.file": "File",
-    "appMenu.file.export": "Export",
+    "appMenu.file.export": "Export...",
     "appMenu.file.import": "Import...",
     "appMenu.file.new": "New...",
     "appMenu.file.open": "Open...",
@@ -1187,7 +1189,13 @@
     "segmentSidebar.slopeStart": "Start Edge (m)",
     "segmentSidebar.slopeEnd": "End Edge (m)",
     "managedStreetSidebar.reloadFromSource": "Reload from {source}",
-    "managedStreetSidebar.reloadConfirm": "Reload this street from {source}? Local segment edits will be lost."
+    "managedStreetSidebar.reloadConfirm": "Reload this street from {source}? Local segment edits will be lost.",
+    "exportModal.title": "Export Scene",
+    "exportModal.glbTitle": "GLB 3D Model",
+    "exportModal.glbDescription": "Download this scene as a .glb 3D model file for use in Blender, Unreal, and other 3D tools.",
+    "exportModal.arReadyLabel": "AR Ready: omit people, vehicles and geospatial layers for augmented reality apps",
+    "exportModal.jsonDescription": "Download this scene as a .3dstreet.json file to back up your work or import into another 3DStreet account.",
+    "exportModal.download": "Download"
   },
   "pt-BR": {
     "actionBar.handTool": "Hand Tool (h) - pan and rotate the view without selecting objects",
@@ -1198,13 +1206,11 @@
     "actionBar.zoomOut": "Zoom Out",
     "addLayer.onlyFirstFileAdded": "Only the first file was added. Drop one file at a time.",
     "addLayer.unsupportedFileType": "Unsupported file type: {fileName}. Supported formats: GLB, GLTF, JPG, PNG, WebP, AVIF, PLY, SPLAT, SPZ.",
-    "appMenu.export.arReadyGlb": "AR Ready GLB",
-    "appMenu.export.glb": "GLB glTF",
     "appMenu.export.gltfError": "Error while trying to save glTF file. Error: {error}",
     "appMenu.export.gltfSuccess": "3DStreet scene exported as glTF file.",
     "appMenu.export.uvTransformSkipped": "UV transformation skipped - using original export",
     "appMenu.file": "File",
-    "appMenu.file.export": "Export",
+    "appMenu.file.export": "Export...",
     "appMenu.file.import": "Import...",
     "appMenu.file.new": "New...",
     "appMenu.file.open": "Open...",
@@ -1782,6 +1788,12 @@
     "segmentSidebar.slopeStart": "Start Edge (m)",
     "segmentSidebar.slopeEnd": "End Edge (m)",
     "managedStreetSidebar.reloadFromSource": "Reload from {source}",
-    "managedStreetSidebar.reloadConfirm": "Reload this street from {source}? Local segment edits will be lost."
+    "managedStreetSidebar.reloadConfirm": "Reload this street from {source}? Local segment edits will be lost.",
+    "exportModal.title": "Export Scene",
+    "exportModal.glbTitle": "GLB 3D Model",
+    "exportModal.glbDescription": "Download this scene as a .glb 3D model file for use in Blender, Unreal, and other 3D tools.",
+    "exportModal.arReadyLabel": "AR Ready: omit people, vehicles and geospatial layers for augmented reality apps",
+    "exportModal.jsonDescription": "Download this scene as a .3dstreet.json file to back up your work or import into another 3DStreet account.",
+    "exportModal.download": "Download"
   }
 }
