+++ src/editor/i18n/locales/en.json
@@ -185,14 +185,8 @@
   "aiChatPanel.inputPlaceholder": {
     "defaultMessage": "Type a command or type /help for options"
   },
-  "appMenu.export.arReadyGlb": {
-    "defaultMessage": "AR Ready GLB"
-  },
   "appMenu.export.exportingGlb": {
     "defaultMessage": "Exporting scene as GLB file..."
-  },
-  "appMenu.export.glb": {
-    "defaultMessage": "GLB glTF"
   },
   "appMenu.export.gltfError": {
     "defaultMessage": "Error while trying to save glTF file. Error: {error}"
@@ -207,7 +201,7 @@
     "defaultMessage": "File"
   },
   "appMenu.file.export": {
-    "defaultMessage": "Export"
+    "defaultMessage": "Export..."
   },
   "appMenu.file.import": {
     "defaultMessage": "Import..."
@@ -895,6 +889,24 @@
   },
   "entity.toggleVisibility": {
     "defaultMessage": "Toggle entity visibility"
+  },
+  "exportModal.arReadyLabel": {
+    "defaultMessage": "AR Ready: omit people, vehicles and geospatial layers for augmented reality apps"
+  },
+  "exportModal.download": {
+    "defaultMessage": "Download"
+  },
+  "exportModal.glbDescription": {
+    "defaultMessage": "Download this scene as a .glb 3D model file for use in Blender, Unreal, and other 3D tools."
+  },
+  "exportModal.glbTitle": {
+    "defaultMessage": "GLB 3D Model"
+  },
+  "exportModal.jsonDescription": {
+    "defaultMessage": "Download this scene as a .3dstreet.json file to back up your work or import into another 3DStreet account."
+  },
+  "exportModal.title": {
+    "defaultMessage": "Export Scene"
   },
   "geoModal.centerpoint": {
     "defaultMessage": "Centerpoint"
