+++ src/editor/i18n/locales/es.json
@@ -61,14 +61,12 @@
   "addLayer.tab.vehicles": "🚗 Vehículos",
   "addLayer.unsupportedFileType": "Tipo de archivo no compatible: {fileName}. Formatos admitidos: GLB, GLTF, JPG, PNG, WebP, AVIF, PLY, SPLAT, SPZ.",
   "aiChatPanel.inputPlaceholder": "Escribe un comando o escribe /help para ver las opciones",
-  "appMenu.export.arReadyGlb": "GLB listo para AR",
   "appMenu.export.exportingGlb": "Exportando escena como archivo GLB...",
-  "appMenu.export.glb": "GLB glTF",
   "appMenu.export.gltfError": "Error al intentar guardar el archivo glTF. Error: {error}",
   "appMenu.export.gltfSuccess": "Escena de 3DStreet exportada como archivo glTF.",
   "appMenu.export.uvTransformSkipped": "Transformación UV omitida: se usa la exportación original",
   "appMenu.file": "Archivo",
-  "appMenu.file.export": "Exportar",
+  "appMenu.file.export": "Exportar...",
   "appMenu.file.import": "Importar...",
   "appMenu.file.new": "Nuevo...",
   "appMenu.file.open": "Abrir...",
@@ -298,6 +296,12 @@
   "entity.dragToReorder": "Arrastra para reordenar",
   "entity.removeEntity": "Eliminar entidad",
   "entity.toggleVisibility": "Alternar visibilidad de la entidad",
+  "exportModal.arReadyLabel": "Listo para AR: omite personas, vehículos y capas geoespaciales para aplicaciones de realidad aumentada",
+  "exportModal.download": "Descargar",
+  "exportModal.glbDescription": "Descarga esta escena como un archivo de modelo 3D .glb para usar en Blender, Unreal y otras herramientas 3D.",
+  "exportModal.glbTitle": "Modelo 3D GLB",
+  "exportModal.jsonDescription": "Descarga esta escena como un archivo .3dstreet.json para respaldar tu trabajo o importarla en otra cuenta de 3DStreet.",
+  "exportModal.title": "Exportar escena",
   "geoModal.centerpoint": "Punto central",
   "geoModal.elevationLabel": "Elevación:",
   "geoModal.freeDuringBeta": "Gratis durante la beta",
