+++ src/editor/i18n/locales/fr.json
@@ -61,14 +61,12 @@
   "addLayer.tab.vehicles": "🚗 Véhicules",
   "addLayer.unsupportedFileType": "Type de fichier non pris en charge : {fileName}. Formats pris en charge : GLB, GLTF, JPG, PNG, WebP, AVIF, PLY, SPLAT, SPZ.",
   "aiChatPanel.inputPlaceholder": "Tapez une commande ou tapez /help pour voir les options",
-  "appMenu.export.arReadyGlb": "GLB prêt pour l'AR",
   "appMenu.export.exportingGlb": "Exportation de la scène en fichier GLB...",
-  "appMenu.export.glb": "GLB glTF",
   "appMenu.export.gltfError": "Erreur lors de l'enregistrement du fichier glTF. Erreur : {error}",
   "appMenu.export.gltfSuccess": "Scène 3DStreet exportée en fichier glTF.",
   "appMenu.export.uvTransformSkipped": "Transformation UV ignorée - utilisation de l'export d'origine",
   "appMenu.file": "Fichier",
-  "appMenu.file.export": "Exporter",
+  "appMenu.file.export": "Exporter...",
   "appMenu.file.import": "Importer...",
   "appMenu.file.new": "Nouveau...",
   "appMenu.file.open": "Ouvrir...",
@@ -298,6 +296,12 @@
   "entity.dragToReorder": "Glisser pour réorganiser",
   "entity.removeEntity": "Supprimer l'entité",
   "entity.toggleVisibility": "Basculer la visibilité de l'entité",
+  "exportModal.arReadyLabel": "Prêt pour l'AR : omet les personnes, les véhicules et les couches géospatiales pour les applications de réalité augmentée",
+  "exportModal.download": "Télécharger",
+  "exportModal.glbDescription": "Télécharger cette scène sous forme de fichier de modèle 3D .glb pour l'utiliser dans Blender, Unreal et d'autres outils 3D.",
+  "exportModal.glbTitle": "Modèle 3D GLB",
+  "exportModal.jsonDescription": "Télécharger cette scène sous forme de fichier .3dstreet.json pour sauvegarder votre travail ou l'importer dans un autre compte 3DStreet.",
+  "exportModal.title": "Exporter la scène",
   "geoModal.centerpoint": "Point central",
   "geoModal.elevationLabel": "Altitude :",
   "geoModal.freeDuringBeta": "Gratuit pendant la bêta",
