+++ src/editor/i18n/locales/pt-BR.json
@@ -61,14 +61,12 @@
   "addLayer.tab.vehicles": "🚗 Veículos",
   "addLayer.unsupportedFileType": "Tipo de arquivo não suportado: {fileName}. Formatos suportados: GLB, GLTF, JPG, PNG, WebP, AVIF, PLY, SPLAT, SPZ.",
   "aiChatPanel.inputPlaceholder": "Digite um comando ou digite /help para ver as opções",
-  "appMenu.export.arReadyGlb": "GLB pronto para AR",
   "appMenu.export.exportingGlb": "Exportando cena como arquivo GLB...",
-  "appMenu.export.glb": "GLB glTF",
   "appMenu.export.gltfError": "Erro ao tentar salvar o arquivo glTF. Erro: {error}",
   "appMenu.export.gltfSuccess": "Cena do 3DStreet exportada como arquivo glTF.",
   "appMenu.export.uvTransformSkipped": "Transformação UV ignorada - usando a exportação original",
   "appMenu.file": "Arquivo",
-  "appMenu.file.export": "Exportar",
+  "appMenu.file.export": "Exportar...",
   "appMenu.file.import": "Importar...",
   "appMenu.file.new": "Novo...",
   "appMenu.file.open": "Abrir...",
@@ -298,6 +296,12 @@
   "entity.dragToReorder": "Arraste para reordenar",
   "entity.removeEntity": "Remover entidade",
   "entity.toggleVisibility": "Alternar visibilidade da entidade",
+  "exportModal.arReadyLabel": "Pronto para AR: omite pessoas, veículos e camadas geoespaciais para aplicativos de realidade aumentada",
+  "exportModal.download": "Baixar",
+  "exportModal.glbDescription": "Baixe esta cena como um arquivo de modelo 3D .glb para usar no Blender, Unreal e outras ferramentas 3D.",
+  "exportModal.glbTitle": "Modelo 3D GLB",
+  "exportModal.jsonDescription": "Baixe esta cena como um arquivo .3dstreet.json para fazer backup do seu trabalho ou importá-la em outra conta do 3DStreet.",
+  "exportModal.title": "Exportar cena",
   "geoModal.centerpoint": "Ponto central",
   "geoModal.elevationLabel": "Elevação:",
   "geoModal.freeDuringBeta": "Grátis Durante o Beta",
